added PreconfiguredSession; when valid, always set environ['REMOTE_USER']
PreconfiguredSession just echoes session data that was given in the initializer.
in server.py we will use it to avoid asking credentials to the user,
when no credentials are specified in the login file.

so far, environ['REMOTE_USER'] was not set when username was not set or empty string
now we can depend on environ['REMOTE_USER'] being set in a valid session
(even when it is a PreconfiguredSession, even when username is empty string)