added fade.c to repository
this is in a very rough state --
it is based on /opt/vc/src/hello_pi/hello_dispmanx/hello_dispmanx.c
with a lot of changes.

this is probably very specific for the raspberry-pi;
the viewer.py code should be changed, to only try to use fading
on raspberry pi -- without fading, also some sleeps can go;
it would be nice to somehow hide this all -- including the additional
sleeps that we use when we are able to use fading --  in the Fading class.