changed naming (fader->shutter, fade in vs out)
this includes changes in the textual commands that form
the interface between the python script and the shutter (fade) program

background:
the reason for this change is the idea that 'shutter' might to be a better
name for a 'device' that hides the transition, from display of one asset,
to the display of the next one.

moreover, fading might just be one particular way to hide the transition.