use right proto (http or https) for splash page
both in server and in viewer -- for this, viewer has to know
what server decides to use, i.e. that decision logic is duplicated;
should be factored out.

use of http vs https is also present in dead code of dbisnewer_ask_server
(changing this code might give a conflict when we remove the dead code
 on the master branch, and then merge the master branch into this one)