after starting asset, give it time to start, and only then start work for next asset
this reduces concurrency at a relatively busy moment:
the moment that the asset has been given the start signal.
At that moment, if things have worked as intented,
that asset has already been prefetched, i.e.
 - the video has already loaded, and is waiting in 'paused' mode, or
 - the web-page or image has already been loaded in the web-browser
   instance that is currently iconified
And thus, at tht moment,
 - the video player has to start playing ('un-pause'), or
 - the iconified web-browser has to be made visible.
This takes already a lot of computing power.

Therefore, it is better to slighly wait,
(until things have quieted down a bit)
before we start work (the prefetch activities) for the next asset
(i.e. before we ask the scheduler for the next asset).