addded fade-to-color and hard-to-color commands
these take a RGB888 color value that is parsed via strtol
and then converted to RGB565 color that is used for drawing.

the idea is that we can use this if we want other fade colors
than black and white, or even, for per-asset fade colors,
once we hace per-asset fade color.

TODO?
make also timing (how fast do we fade) a command parameter?